requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
alembic
python-multipart
email-validator
itsdangerous
